Deduce the page size and use that in calculations. (#447)
Pages are allocated by page order number. Each page fragment
is described by page_offset and size. The page order for a
page that can hold the fragment is calculated as get_order()
of <page_offset> + <fragment size> bytes. As other fragments
in the same page are taken into account, the calculated page
order is corrected. The page order leads to the page size.
In the end, if the page's reference count is one, then the
deduced page size is the correct size (because there're no
more fragments in the page), and the room has been found in
that page.